first task completed (without numpy and work with names of files)
